:sparkles: useFunnel 훅으로 컴포넌트 로직 분리
