--- src/components/goal/Step.tsx
@@ -1,13 +1,0 @@
-type Props = {
-  check: boolean;
-  children: React.ReactNode;
-};
-
-const Step = ({ check, children }: Props) => {
-  if (check === true) {
-    return children;
-  }
-  return null;
-};
-
-export default Step;
