--- src/constants/step.ts
@@ -1,1 +1,0 @@
-export const steps = ['seed','routines','duration']
